VHDtool: properly handle creating/converting 127GiB images.
The '-c' option allows generating (whether converting or creating)
VHD images that, when used with the emulated IDE controller
in Hyper-V, will not make the disk appear smaller than
the image size.

This matters if you want VHD images that will work correctly
when moved from virtual SCSI to IDE, or when converting
existing raw images into VHDs to be used, again, with the
emulated IDE controller.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>